+++ infra/terraform/providers.tf
@@ -6,9 +6,9 @@ terraform {
       source  = "hashicorp/azurerm"
       version = "~> 4.2"
     }
-    null = {
-      source  = "hashicorp/null"
-      version = "~> 3.0"
+    azapi = {
+      source  = "azure/azapi"
+      version = "~> 2.0"
     }
   }
 
@@ -22,3 +22,6 @@ provider "azurerm" {
   storage_use_azuread = true
   features {}
 }
+
+provider "azapi" {}
+